requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.74
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# LLM integrations
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.8.0

# Optional: For future database integration
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# redis>=5.0.0 